NOTE: PHP >5.4 now required as of this version. (May be temporary)
Found out today that the Socket.php is a PHP Stream not a PHP Socket, this still is the case now, but may change later because:

Re-implemented the Bot.php code to use socket_select over the local socket and the IRC socket, adding the workhorse socket whenit it gets one.
By using socket_select I have reduced the CPU (on a test machine) from 10% to 0% (average) due to both sockets now being blocking again, and the tight loop now being less tight because of that.
PHP 5.4 is now needed due to socket_import_stream being used to get the socket from the PHP Stream mentioned above.
I may try to re-write Socket.php to use a real socket and therefore remove this version constraint, although it could be possible that I use other PHP 5.4 features in the future such as shot array syntax.